Полностью переписал модели запросов и ответов в постах

+++ TgPics.WebApi/Services/PostsService.cs
@@ -8,10 +8,10 @@ using TgPics.WebApi.Models;
 
 public interface IPostsService
 {
-    public void Add(AddPostRequest post);
-    public void Remove(Post post);
-    public List<Post> GetAll();
-    public void RemoveAll();
+    public void Add(AddPostRequest request);
+    public void Remove(RemovePostRequest request);
+    public GetAllPostsResponse GetAll();
+    public void RemoveAll(RemoveAllPostsRequest request);
 }
 
 public class PostsService : IPostsService
@@ -21,31 +21,33 @@ public class PostsService : IPostsService
         this.settings = settings.Value;
     }
 
+    private const string CONFIRMATION = "yeah, kill 'em all.";
+
     private readonly AppSettings settings;
 
-    public void Add(AddPostRequest post)
+    public void Add(AddPostRequest request)
     {
         using var database = new DBService();
 
-        var postEntity = new Post
+        var post = new Post
         {
-            SourceLink = post.SourceLink,
-            SourceTitle = post.SourceTitle,
-            Comment = post.Comment,
-            Pictures = post.Pictures.Select(p => new Picture
+            SourceLink = request.SourceLink,
+            SourceTitle = request.SourceTitle,
+            Comment = request.Comment,
+            Pictures = request.Pictures.Select(p => new Picture
             {
                 Url = p.Url,
                 Position = p.Position
             })
         };
 
-        foreach (var pic in postEntity.Pictures)
+        foreach (var pic in post.Pictures)
         {
             if (pic.Data == null && pic.Url != null)
                 pic.Load();
         }
 
-        if (postEntity.PublicationDateTime == null)
+        if (post.PublicationDateTime == null)
         {
             if (database.Posts.Any())
             {
@@ -54,46 +56,62 @@ public class PostsService : IPostsService
 
                 var timing = lastPost.PublicationDateTime?.TimeOfDay;
                 var nextTiming = settings.Timings.First(t => t > timing);
-                
-                postEntity.PublicationDateTime = lastPost
+
+                post.PublicationDateTime = lastPost
                     .PublicationDateTime?.Date.Add(nextTiming);
             }
             else
             {
                 var timing = DateTime.Now.TimeOfDay;
                 var nextTiming = settings.Timings.First(t => t > timing);
-                postEntity.PublicationDateTime = DateTime.Now.Date.Add(nextTiming);
+                post.PublicationDateTime = DateTime.Now.Date.Add(nextTiming);
             }
         }
 
-        database.Posts.Add(postEntity);
+        database.Posts.Add(post);
         database.SaveChanges();
     }
 
-    public void Remove(Post post)
+    public void Remove(RemovePostRequest request)
     {
         using var database = new DBService();
-        if (database.Posts.Contains(post))
+
+        var post = database.Posts
+            .Include(p => p.Pictures)
+            .FirstOrDefault(p => p.Id == request.PostId);
+
+        if (post != null)
         {
             database.Posts.Remove(post);
             database.SaveChanges();
         }
+        else throw new ArgumentException(
+            "A post with this id was not found.", nameof(request));
     }
 
-    public List<Post> GetAll()
+    public GetAllPostsResponse GetAll()
     {
         using var database = new DBService();
-        return database.Posts
-            .AsNoTracking()
-            .Include(p => p.Pictures)
-            .ToList();
+        return new GetAllPostsResponse
+        {
+            Posts = database.Posts
+                .AsNoTracking()
+                .Include(p => p.Pictures)
+                .ToList()
+        };
     }
 
-    public void RemoveAll()
+    public void RemoveAll(RemoveAllPostsRequest request)
     {
-        using var database = new DBService();
-        database.Posts.RemoveRange(
-            database.Posts.Include(p => p.Pictures));
-        database.SaveChanges();
+        if (request.Confirmation == CONFIRMATION)
+        {
+            using var database = new DBService();
+            database.Posts.RemoveRange(
+                database.Posts.Include(p => p.Pictures));
+            database.SaveChanges();
+        }
+        else throw new ArgumentException(
+            $"To confirm the removal, you should pass Confirmation='{CONFIRMATION}'",
+            nameof(request));
     }
 }
